requirements.txt
telethon>=1.34.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
